+++ libs/cardano-ledger-core/src/Cardano/Ledger/Metadata.hs
@@ -22,14 +22,14 @@ import Cardano.Ledger.Binary (
   TokenType (..),
   cborError,
   decodeBreakOr,
-  decodeBytesIndef,
+  decodeByteArrayIndefLen,
   decodeInteger,
   decodeListLen,
   decodeListLenIndef,
   decodeMapLen,
   decodeMapLenIndef,
   decodeString,
-  decodeStringIndef,
+  decodeStringIndefLen,
   encodeInteger,
   encodeListLen,
   encodeMapLen,
@@ -37,6 +37,7 @@ import Cardano.Ledger.Binary (
   getDecoderVersion,
   natVersion,
   peekTokenType,
+  unBA,
  )
 import Cardano.Ledger.Orphans ()
 import Control.Applicative ((<|>))
@@ -168,8 +169,7 @@ decodeMetadatum = do
         decodeError "bytes .size (0..64): bytestring exceeds 64 bytes"
       return (B ba)
     TypeBytesIndef -> do
-      decodeBytesIndef
-      !ba <- decodeBytesIndefLen []
+      !ba <- unBA <$> decodeByteArrayIndefLen
       when (checkSizes && Prim.sizeofByteArray ba > 64) $
         decodeError "bytes .size (0..64): bytestring exceeds 64 bytes"
       return (B ba)
@@ -179,8 +179,7 @@ decodeMetadatum = do
         decodeError "text .size (0..64): text exceeds 64 bytes"
       return (S x)
     TypeStringIndef -> do
-      decodeStringIndef
-      !x <- decodeStringIndefLen []
+      !x <- decodeStringIndefLen
       when (checkSizes && TF.lengthWord8 x > 64) $
         decodeError "text .size (0..64): text exceeds 64 bytes"
       return (S x)
@@ -220,24 +219,6 @@ decodeMetadatum = do
     _ -> decodeError ("Unsupported token type " <> T.pack (show tkty))
   where
     decodeError msg = cborError (DecoderErrorCustom "metadata" msg)
-
-decodeBytesIndefLen :: [ByteArray] -> Decoder s ByteArray
-decodeBytesIndefLen acc = do
-  stop <- decodeBreakOr
-  if stop
-    then return $! mconcat $ reverse acc
-    else do
-      !ba <- decCBOR
-      decodeBytesIndefLen (ba : acc)
-
-decodeStringIndefLen :: [T.Text] -> Decoder s T.Text
-decodeStringIndefLen acc = do
-  stop <- decodeBreakOr
-  if stop
-    then return $! T.concat (reverse acc)
-    else do
-      !str <- decodeString
-      decodeStringIndefLen (str : acc)
 
 decodeListN :: Int -> [Metadatum] -> Decoder s [Metadatum]
 decodeListN !n acc =