+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Version history for `cardano-ledger-binary`
 
-## 1.9.0.1
+## 1.10.0.0
 
-*
+* Move `decodeIPv4` and `decodeIPv6` into `Cardano.Ledger.Binary.Decoding.DecCBOR`
+* Add `decodeStringIndefLen`, `decodeStringDefOrIndef`, `decodeBytesIndefLen`, `decodeBytesDefOrIndef`, `decodeByteArrayIndefLen` and `decodeByteArrayDefOrIndef`
+* Change `DecCBOR` instances for `Text` and tuples to accept indefinite-length encoding starting with protocol version 12
 
 ## 1.9.0.0